pyproject.toml
[project]
name = "sla-process"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.3",
    "cupy-cuda13x>=13.6.0",
    "opencv-python>=4.11.0.86",
    "scikit-image>=0.25.2",
    "scipy>=1.16.2",
    "matplotlib>=3.10.6",
    "pythonnet>=3.0.5",
    "tqdm>=4.67.1",
    "line-profiler>=5.0.0",
    "ipython>=9.6.0",
    "ipykernel>=6.30.1",
    "napari[all]>=0.6.6",
    "qtpy>=2.4.3",
    "pyqt5>=5.15.11",
    "rectpack>=0.2.2",
]

[tool.setuptools.packages.find]
where = ["src"]
